# backend/app/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: 3TechMarket API
  description: "3TechMarket API, fancyyyyyy"
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/products/{id}:
    put:
      tags:
      - product-rest-controller
      summary: Update a product by id
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Product updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      tags:
      - product-rest-controller
      summary: Delete a product by id
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                type: object
        "204":
          description: Product deleted
...
  /api/products/price-history/{id}:
    post:
      tags:
      - product-rest-controller
      summary: Get a product's price history
      operationId: getProductPriceHistory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Product's price history retrieved
          content:
            '*/*':
              schema:
                type: object
  /api/products/add-product:
    post:
      tags:
      - product-rest-controller
      summary: Create a new product
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Product not created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
        "201":
          description: Product created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
        "403":
          description: User not authorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
...
  /api/products:
    get:
      tags:
      - product-rest-controller
      summary: Get all products
      operationId: getAllProducts
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Products retrieved
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageProduct'
  /api/products/search/:
    get:
      tags:
      - product-rest-controller
      summary: Search products by name
      operationId: searchProducts
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: product
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Products retrieved
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageProduct'
  /api/products/info/{id}:
    get:
      tags:
      - product-rest-controller
      summary: Get a product by id
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Product retrieved
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Product'
...
components:
  schemas:
    Image:
      type: object
      properties:
        imageId:
          type: integer
          format: int64
        fileName:
          type: string
...
    Product:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        description:
          type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        productPrice:
          type: number
          format: double
        mainImage:
          $ref: '#/components/schemas/Image'
        tags:
          type: array
          items:
            type: string
        productStock:
          type: integer
          format: int32
        productName:
          type: string
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        id:
          type: integer
          format: int64
        productPrices:
          type: array
          items:
            type: number
            format: double
        tagList:
          type: string
    Review:
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        reviewTitle:
          type: string
        rating:
          type: integer
          format: int32
        reviewText:
          type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
...
    PageProduct:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...